pyproject.toml
[project]
name = "reviewhub"
version = "0.1.0"
description = ""
authors = [
    {name = "Sash_Du",email = "sashadubina1996@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "environ (>=1.0,<2.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "django-ninja (>=1.4.1,<2.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "dependency-injector (>=4.46.0,<5.0.0)",
    "punq (>=0.7.0,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
pre-commit = "^3.6.0"
faker = "^22.6.0"
pytest-django = "^4.8.0"
pytest = "^8.3.5"
factory-boy = "^3.3.3"



[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.project.settings"


[tool.isort]
multi_line_output=3
include_trailing_comma=true
line_length=120
lines_after_imports=2
lines_between_sections=1
lines_between_types=0
force_alphabetical_sort_within_sections=true
ensure_newline_before_comments=true
combine_as_imports=true
combine_star=true
skip_gitignore=true
balanced_wrapping=true
force_grid_wrap=2
src_paths=["core", ]
skip_glob=[
    ".git",
    ".idea",
    ".vscode",
    "venv",
    ".venv",
]
known_django=["django","ninja"]
known_first_party=["core"]
sections=[
    "FUTURE",
    "STDLIB",
    "DJANGO",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
